I left only three classes, made everything work, there are bugs.

+++ src/CalculatorProcessor.java
@@ -1,19 +1,19 @@
 
 public class CalculatorProcessor {
 	
-	public static int Sum( int number1, int number2) {
+	public static Double Sum( Double number1, Double number2) {
 	    return number1 + number2; 
 	}
 	
-	public static int Minus(int number1, int number2) {
+	public static Double Minus(Double number1, Double number2) {
 	    return number1 - number2; 
 	    }
 	
-	public static int multiply(int number1, int number2) {
+	public static Double multiply(Double number1, Double number2) {
 	    return number1 * number2; 
 	    }
 	
-	public static int Divide(int number1, int number2) {
+	public static Double Divide(Double number1, Double number2) {
 	    return number1 / number2; 
 	    }
